requirements.txt
asgiref==3.6.0
Django==4.1.4
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
